updated models.py on ireporterApp
