+++ authenticationApp/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from .models import  CustomUser,Intervention
+from .models import  CustomUser
 from django.contrib import auth
 from rest_framework.exceptions import AuthenticationFailed
 
@@ -27,8 +27,5 @@ class LoginSerializer(serializers.ModelSerializer):
     class Meta:
         model = CustomUser
         fields=('email','password','token',)  
-class InterventionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Intervention
-        fields=('subject','description','location',)  
+
